pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dbviewer"
version = "1.0.0"
description = "Web-based database viewer and management tool"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "fastapi>=0.104",
    "uvicorn[standard]>=0.24",
    "bcrypt>=4.0",
    "python-multipart>=0.0.6",
    "pymysql>=1.1",
    "psycopg2-binary>=2.9",
    "pymssql>=2.2",
    "openpyxl>=3.1",
    "httpx>=0.25",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "httpx>=0.25",
]

[project.scripts]
dbviewer = "dbviewer.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
dbviewer = ["static/*"]
